feat: add 'assert_true' macro and refactor 'Invitation' structure
Added a new 'assert_true' macro in 'assertion.rs' for enhancing
capability of doing assertions. Refactored 'Invitation' structure in
'invitation.rs' to handle invitation start and end validity more
flexible. Also, now it checks if start date is before end date whenever
we redefine the validity period.

The refactoring eased the circulation and the maintainability of
invitation dates.
We also added a list of invitations in the 'Tenant' structure to keep
track of all invitations per tenant.